requirements.txt
# Kalshi Edge Detection Engine
# Python 3.10+ required

# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# mypy>=1.5.0
# types-requests>=2.31.0
# types-PyYAML>=6.0.0
